reorganize ServerParams to merge into config_struct.py
